# spec/api.openapi.yaml
openapi: 3.0.3
info:
  title: Lula Coffee API
  version: 1.0.0
  description: |
    Ecommerce API for Lula Coffee (coffee, organic coffee, mugs, tea).
    Contract source of truth for frontend and integrations.

servers:
  - url: /api
    description: Same-origin Next.js route handlers

tags:
  - name: Health
  - name: Catalog
  - name: Customer
  - name: Admin

paths:
  /health:
    get:
      tags: [Health]
      summary: Service health check
      operationId: getHealth
      security: []
      responses:
        '200':
          description: API is reachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/products:
    get:
      tags: [Catalog]
      summary: List published products
      operationId: listProducts
      security: []
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ProductCategory'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 24
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Product page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'

  /v1/products/{productId}:
    get:
      tags: [Catalog]
      summary: Get product by id
      operationId: getProduct
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/me:
    get:
      tags: [Customer]
      summary: Current authenticated customer profile stub
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/cart:
    get:
      tags: [Customer]
      summary: Get cart for authenticated customer
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Customer]
      summary: Replace cart lines (customer role)
      operationId: replaceCart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartReplaceRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'

  /v1/admin/products:
    get:
      tags: [Admin]
      summary: List all products (admin)
      operationId: adminListProducts
      security:
        - bearerAuth: []
      x-required-role: admin
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin]
      summary: Create product (admin)
      operationId: adminCreateProduct
      security:
        - bearerAuth: []
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'

  /v1/admin/products/{productId}:
    patch:
      tags: [Admin]
      summary: Update product (admin)
      operationId: adminUpdateProduct
      security:
        - bearerAuth: []
      x-required-role: admin
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin]
      summary: Delete product (admin)
      operationId: adminDeleteProduct
      security:
        - bearerAuth: []
      x-required-role: admin
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: unauthorized
            message: Bearer token required or invalid
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: forbidden
            message: Insufficient role for this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    HealthResponse:
      type: object
      required: [status, service, timestamp]
      properties:
        status:
          type: string
          enum: [ok]
        service:
          type: string
          example: lula-coffee-api
        timestamp:
          type: string
          format: date-time
        database:
          type: string
          enum: [connected, unavailable]
          description: Best-effort DB probe

    ProductCategory:
      type: string
      enum: [coffee, organic_coffee, mugs, tea]

    Product:
      type: object
      required:
        - id
        - createdAt
        - updatedAt
        - name
        - slug
        - priceCents
        - category
        - sku
        - inStock
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        priceCents:
          type: integer
        category:
          $ref: '#/components/schemas/ProductCategory'
        sku:
          type: string
        imageUrl:
          type: string
          nullable: true
        inStock:
          type: boolean

    ProductResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Product'

    ProductListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pagination:
          $ref: '#/components/schemas/Pagination'

    Pagination:
      type: object
      required: [limit, offset, total]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer

    ProductCreateRequest:
      type: object
      required: [name, slug, priceCents, category, sku]
      properties:
        name:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 220
        description:
          type: string
        priceCents:
          type: integer
          minimum: 0
        category:
          $ref: '#/components/schemas/ProductCategory'
        sku:
          type: string
          maxLength: 64
        imageUrl:
          type: string
        inStock:
          type: boolean
          default: true

    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        priceCents:
          type: integer
        category:
          $ref: '#/components/schemas/ProductCategory'
        sku:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean

    CartLine:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1

    CartReplaceRequest:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/CartLine'

    CartResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [userId, lines]
          properties:
            userId:
              type: string
            lines:
              type: array
              items:
                type: object
                required: [id, createdAt, updatedAt, productId, quantity, product]
                properties:
                  id:
                    type: string
                    format: uuid
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
                  productId:
                    type: string
                    format: uuid
                  quantity:
                    type: integer
                  product:
                    $ref: '#/components/schemas/Product'

    MeResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, role]
          properties:
            id:
              type: string
            role:
              type: string
              enum: [customer, admin]
            email:
              type: string
              format: email

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true